requirements.txt
streamlit
google-generativeai
feedparser
pytz
ccxt
pandas
numpy
yfinance
requests
streamlit-components
streamlit-autorefresh